+++ src/ciphers/index.js
@@ -1,0 +1,14 @@
+const  encryptСaesar = require('./encrypt-caesar'),
+        decodeСaesar = require('./decode-caesar'),
+        cipherAtbash = require('./cipher-atbash'),
+        encryptRot = require('./encrypt-rot'),
+        decodeRot = require('./decode-rot')
+
+
+module.exports = {
+  encryptСaesar,
+  decodeСaesar,
+  cipherAtbash,
+  encryptRot,
+  decodeRot
+};
