+++ src/streams/custom-read-stream.js
@@ -1,12 +1,9 @@
 const fs = require('fs');
-const arrArguments = process.argv.slice(2);
 const { Readable } = require('stream');
-const { Writable } = require('stream');
-let input = '';
-let output = '';
+const arrArguments = process.argv.slice(2);
 
+let input = '';
 arrArguments.indexOf('-i') + 1 ? input = arrArguments[arrArguments.indexOf('-i') + 1] : arrArguments.indexOf('--input') + 1 ? input = arrArguments[arrArguments.indexOf('--input') + 1] : input = './input.txt';
-arrArguments.indexOf('-o') + 1 ? output = arrArguments[arrArguments.indexOf('-o') + 1] : arrArguments.indexOf('--output') + 1 ? output = arrArguments[arrArguments.indexOf('--output') + 1] : output = './output.txt';
 
 
 class ReadStream extends Readable {
@@ -44,38 +41,9 @@ class ReadStream extends Readable {
   }
 }
 
-class WriteStream extends Writable {
-  constructor(filename) {
-    super();
-    this.filename = filename;
-  }
-  _construct(callback) {
-    fs.open(this.filename, 'a', (err, fd) => {
-      if (err) {
-        callback(err);
-      } else {
-        this.fd = fd;
-        callback();
-      }
-    });
-  }
-  _write(chunk, encoding, callback) {
-    fs.write(this.fd, chunk, callback);
-  }
-  _destroy(err, callback) {
-    if (this.fd) {
-      fs.close(this.fd, (er) => callback(er || err));
-    } else {
-      callback(err);
-    }
-  }
-}
-
 const readStream = new ReadStream(input);
-const writeStream = new WriteStream(output);
 
 
 module.exports = {
-  readStream,
-  writeStream
+  readStream
 };
