+++ src/streams/index.js
@@ -1,0 +1,8 @@
+const {readStream} = require('./custom-read-stream');
+const {writeStream} = require('./custom-write-stream');
+
+
+module.exports = {
+  readStream,
+  writeStream
+}
