+++ src/validation/index.js
@@ -1,0 +1,11 @@
+const validationRepeatedArguments = require('./validation-repeated-arguments'),
+      validationConfig = require('./validation-config'),
+      validationInput = require('./validation-input'),
+      validationOutput = require('./validation-output')
+
+module.exports = {
+  validationRepeatedArguments,
+  validationConfig,
+  validationInput,
+  validationOutput
+};
